+++ tests/test_plotting.py
@@ -3,7 +3,7 @@
 import gym
 import numpy as np
 import bauwerk
-import bauwerk.utils.plotting
+import bauwerk.utils.env_plotting
 import bauwerk.utils.data
 import bauwerk.utils.gym
 
@@ -12,30 +12,35 @@ def test_core_plotting():
 
     env = gym.make("bauwerk/House-v0")
     initial_obs = bauwerk.utils.gym.force_old_reset(env.reset())
-    _ = bauwerk.utils.plotting.EnvPlotter(
-        initial_obs=initial_obs,
+    _ = bauwerk.utils.env_plotting.EnvPlotter(
         env=env,
         include_house_figure=True,
     )
-    _ = bauwerk.utils.plotting.EnvPlotter(
-        initial_obs=initial_obs,
+    _ = bauwerk.utils.env_plotting.EnvPlotter(
         env=env,
     )
-    _ = bauwerk.utils.plotting.EnvPlotter(
-        initial_obs=initial_obs, env=env, alternative_plotting=False
-    )
+    _ = bauwerk.utils.env_plotting.EnvPlotter(env=env, alternative_plotting=False)
 
 
 def test_opt_acts_plotting():
 
     env = gym.make("bauwerk/House-v0")
-    bauwerk.utils.plotting.plot_optimal_actions(env, max_num_acts=365)
+    bauwerk.utils.env_plotting.plot_optimal_actions(env, max_num_acts=365)
 
 
 def test_x_axis():
     env = gym.make("bauwerk/House-v0")
-    initial_obs = bauwerk.utils.gym.force_old_reset(env.reset())
-    plotter = bauwerk.utils.plotting.EnvPlotter(
-        initial_obs=initial_obs, env=env, alternative_plotting=True
-    )
+    env.reset()
+    plotter = bauwerk.utils.env_plotting.EnvPlotter(env=env, alternative_plotting=True)
     assert np.allclose(plotter.line_x[:4], np.array([0, 1, 2, 3]))
+
+
+def test_builtin_plotting():
+    # add render_mode to config to enable plotting
+    env = gym.make("bauwerk/House-v0", cfg={"render_mode": "rgb_array"})
+    env.reset()
+    for _ in range(24):
+        action = env.action_space.sample()
+        env.step(action)
+
+    env.render()